Add cf7Input file
