+++ packages/contact-form-7/src/Input.js
@@ -1,4 +1,5 @@
 import { FormIdContext } from "./Form";
+import { connect } from "frontity";
 
 const Input = ({ state, actions, name }) => {
 	const id = React.useContext(FormIdContext);
@@ -10,3 +11,5 @@ const Input = ({ state, actions, name }) => {
 		<input onChange={onChange}/>
 );
 };
+
+export default connect( Input );
